### 2024-06-08 #6858 @editorialbot
[PRE REVIEW]: GWSurrogate: A Python package for gravitational wave surrogate models
**Submitting author:** @sfield17 (<a href="http://orcid.org/0000-0002-6037-3277">Scott Field</a>)
**Repository:** https://github.com/sxs-collaboration/gwsurrogate
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v1.1.6
