+++ tutorial/paper/paper.md
@@ -49,7 +49,7 @@ affiliations:
     index: 2
   - name: Institute for Gravitational and Subatomic Physics (GRASP), Utrecht University, 3584 CC Utrecht, The Netherlands
     index: 3
-  - name: Jet Propulsion Laboratory, California Institute of Technology, Pasadena, CA 91109, USA
+  - name: Theoretical Astrophysics, Walter Burke Institute for Theoretical Physics, California Institute of Technology, Pasadena, California 91125, USA
     index: 4
   - name: Kavli Institute for Theoretical Physics, University of California Santa Barbara, CA 93106, USA
     index: 5
@@ -104,6 +104,6 @@ Currently, there are 15 supported surrogate models (@Blackman:2015pia, @OShaughn
 
 # Acknowledgements
 
-We acknowledge our many close collaborators for their contribution to the development of surrogate models. We further acknowledge the community of ``GWSurrogate`` users who have contributed pull requests and opened issues, including Kevin Barkett, Mike Boyle, Collin Capano, Dwyer Deighan, Raffi Enficiaud, Oliver Jennrich, Gaurav Khanna, Duncan Macleod, Alex Nitz, Seth Olsen, Swati Singh, and Avi Vajpeyi. ``GWSurrogate`` has been developed over the past 10 years with continued support from the National Science Foundation, most recently through NSF grants PHY-2110496, PHY-2309301, and DMS-2309609. This work was also supported in part by the Sherman Fairchild Foundation, by NSF Grants PHY-2207342 and OAC-2209655 at Cornell, and by NSF Grants PHY-2309211, PHY-2309231, and OAC-2209656 at Caltech. The research conducted by C.R.G. was carried out partially at the Jet Propulsion Laboratory, California Institute of Technology, under a contract with the National Aeronautics and Space Administration (80NM0018D0004).
+We acknowledge our many close collaborators for their contribution to the development of surrogate models. We further acknowledge the community of ``GWSurrogate`` users who have contributed pull requests and opened issues, including Kevin Barkett, Mike Boyle, Collin Capano, Dwyer Deighan, Raffi Enficiaud, Oliver Jennrich, Gaurav Khanna, Duncan Macleod, Alex Nitz, Seth Olsen, Swati Singh, and Avi Vajpeyi. ``GWSurrogate`` has been developed over the past 10 years with continued support from the National Science Foundation, most recently through NSF grants PHY-2110496, PHY-2309301, and DMS-2309609. This work was also supported in part by the Sherman Fairchild Foundation, by NSF Grants PHY-2207342 and OAC-2209655 at Cornell, and by NSF Grants PHY-2309211, PHY-2309231, and OAC-2209656 at Caltech. 
 
 # References
