### 2024-07-11 c950386
Treat precessing surrogate fit parameters with model level scope.
Previously, the function that converted the users parameter inputs
to the fit's parameterization was global in scope and hardcoded
to NRSur7dqq4.

This change moves the function so that its specific to each model.
