### 2024-07-11 83babd2
Treat precessing surrogate fit settings with model level scope.
Previously, the function that stored information about each
model's fit settings was global in scope and hardcoded
to NRSur7dqq4.

This change moves the function so that its specific to each model.
