### 2024-07-12 62013a8
Deep copy input dictionaries to avoid local modification.
Dictionary inputs were processed by pop, modifying the users
input. This fix does a deep copy on the input dicts to avoid that.
