### 2024-07-12 5a3553b
Move more model specific settings out of PrecessingSurrogate classes
* Model-specific ellMax and largest allowable orbitical frequency
were previously hardcoded. These model setting parameters are now
passed and set when creating precessing surrogate model objects.
* Allow precessing surrogate model objects to be either created
by passing an hdf5 file name or hdf5 file groups. The former
was already supported. The latter can be useful in some context,
such as multi-domain surrogate models.
