### 2024-07-16 #6858 @sfield17
Hi Paul,

 Here are a few potential reviewers. All of these folks are involved in
gravitational wave science and have contributed to various public codes
used in our community. Many are also experts in gravitational wave modeling.

Patricia Schmidt
Ian Harry
Hector Estelles
Cecilio Garcia-Quiros
Duncan Macleod
Raffi Enficiaud

 I've never participated in a JOSS review before, so I don't know how this
works. For contact purposes, do you need their email addresses or GitHub
IDs?

Scott

On Mon, Jul 15, 2024 at 5:20 PM Paul La Plante ***@***.***>
wrote:

> @sfield17 <https://github.com/sfield17> — Thanks for your patience! I
> will be the editor for this submission. I will begin the review process by
> reaching out to potential reviewers. Once two have agreed to review this
> submission, we can begin the actual review process.
>
> If you know of anyone who might be suitable, please feel free to let me
> know and I can reach out to them. On my end, I will add reviewers to this
> issue as they agree. Let me know if you have any questions!
>
