### 2024-07-28 sxs-collaboration/gwsurrogate#51 @ColmTalbot
Incompatible with numpy v2
Currently gwsurrogate fails with numpy v2.

The first error I saw is due to the use of `np.string_` which has been removed. There is a migration guide at https://numpy.org/devdocs/numpy_2_0_migration_guide.html that describes an automated check that can be run.
