### 2024-08-21 384b52a
Support for multi-domain precessing surrogate model SEOBNRv4PHMSur.
* Add SEOBNRv4PHMSur as an entry in the surrogate catalog
* New class PrecessingSurrogateMultiDomain for precessing surrogate models
  defined on multiple subdomains over the parameter space
* New surrogate evaluator class SEOBNRv4PHMSur
* Generalizations of precessing and dynamics surrogate classes
  to handle more generic models.
