+++ gwsurrogate/new/precessing_surrogate.py
@@ -714,21 +714,29 @@ class CoorbitalWaveformSurrogate:
         self.data = {}
         self.mode_list = []
         for ell in range(2, self.ellMax+1):
-            # m=0 is different
-            self.mode_list.append( (ell,0) )
-            for reim in ['real', 'imag']:
-                group = h5file['hCoorb_%s_0_%s'%(ell, reim)]
-                self.data['%s_0_%s'%(ell, reim)] \
-                        = _extract_component_data(group)
+            # m=0 has a different file naming convention
+            # If there are no modes, skip. Assumes if group "*_real" exists, then 
+            # "*_imag" exists as well.
+            if self._check_h5group_exists(h5file,'hCoorb_%s_0_real'%ell): 
+                #print("Loading (ell=%s,0) mode"%(ell))
+                self.mode_list.append( (ell,0) )
+                for reim in ['real', 'imag']:
+                    group = h5file['hCoorb_%s_0_%s'%(ell, reim)]
+                    self.data['%s_0_%s'%(ell, reim)] \
+                            = _extract_component_data(group)
 
             for m in range(1, ell+1):
-                self.mode_list.append( (ell,m) )
-                self.mode_list.append( (ell,-m) )
-                for reim in ['Re', 'Im']:
-                    for pm in ['+', '-']:
-                        group = h5file['hCoorb_%s_%s_%s%s'%(ell, m, reim, pm)]
-                        tmp_data = _extract_component_data(group)
-                        self.data['%s_%s_%s%s'%(ell, m, reim, pm)] = tmp_data
+                # If there are no modes, skip. Assumes if group "*_Re+" exists, then
+                # "*_Re-", "*_Im+", and "_Im-" exists as well.
+                if self._check_h5group_exists(h5file,'hCoorb_%s_%s_Re+'%(ell, m)): 
+                    #print("Loading (ell=%s,m=\pm%s) modes"%(ell, m))
+                    self.mode_list.append( (ell,m) )
+                    self.mode_list.append( (ell,-m) )
+                    for reim in ['Re', 'Im']:
+                        for pm in ['+', '-']:
+                            group = h5file['hCoorb_%s_%s_%s%s'%(ell, m, reim, pm)]
+                            tmp_data = _extract_component_data(group)
+                            self.data['%s_%s_%s%s'%(ell, m, reim, pm)] = tmp_data
 
 
     def __call__(self, q, chiA, chiB, ellMax=4):
@@ -743,19 +751,30 @@ ellMax: The maximum ell mode to evaluate.
         modes = 1.j*np.zeros((nmodes, len(self.t)))
 
         for ell in range(2, ellMax+1):
-            # m=0 is different
-            re = self._eval_comp(self.data['%s_0_real'%(ell)], q, chiA, chiB)
-            im = self._eval_comp(self.data['%s_0_imag'%(ell)], q, chiA, chiB)
-            modes[ell*(ell+1) - 4] = re + 1.j*im
+            
+            # NOTE: modes is populated with zeros, so skipping means 
+            #       "set to zero". This is required for inertial_waveform_modes,
+            #       which maps from coorb to inertial frame modes. 
+            # m=0 has a different evaluation pattern
+            if (ell,0) in self.mode_list:
+                re = self._eval_comp(self.data['%s_0_real'%(ell)], q, chiA, chiB)
+                im = self._eval_comp(self.data['%s_0_imag'%(ell)], q, chiA, chiB)
+                modes[ell*(ell+1) - 4] = re + 1.j*im
+                #print("evaluation ell=%s, m=0"%ell)
 
+            # NOTE: modes is populated with zeros, so skipping means 
+            # "set to zero". This is required for inertial_waveform_modes,
+            # which maps from coorb to inertial frame modes. 
             for m in range(1, ell+1):
[…]
