+++ gwsurrogate/surrogate.py
@@ -2485,8 +2485,9 @@ See the __call__ method on how to evaluate waveforms.
             chi_max_bfOrder = 2
             return q_fit_offset, q_fit_slope, q_max_bfOrder, chi_max_bfOrder
 
-        # largest ell in the surrogate
+        # largest ell mode in the surrogate
         ellMax_NRSur7dq4 = 4
+
         # max allowable reference dimensionless orbital angular frequency
         omega_ref_max_NRSur7dq4 = 0.201
 
@@ -2505,6 +2506,177 @@ See the __call__ method on how to evaluate waveforms.
         x = [q, chiA0, chiB0]
         return x
 
+class SEOBNRv4PHMSur(SurrogateEvaluator):
+    """
+A class for the SEOBNRv4PHM surrogate models.
+
+Evaluates gravitational waveforms generated by precessing binary black hole
+systems with generic mass ratios and spins.
+
+This model includes the following spin-weighted spherical harmonic modes:
+2<=ell<=5, -ell<=m<=ell in inertial frame.
+
+The parameter space of validity is:
+q \in [1, 20], and |chi1|,|chi2| \in [-0.99, 0.99], with generic directions.
+where q is the mass ratio and chi1/chi2 are the spin vectors of the
+heavier/lighter BH, respectively.
+
+The surrogate has been trained in the range
+q \in [1, 20] and |chi1|/|chi2| \in [-0.8, 0.8], but produces reasonable
+waveforms in the above range and has been tested against existing
+NR waveforms in that range.
+
+See the __call__ method on how to evaluate waveforms.
+In the __call__ method, x must have format x = [q, chi1, chi2].
+    """
+
+    def __init__(self, h5filename):
+        self.h5filename = h5filename
+        domain_type = 'Time'
+        keywords = {
+            'Precessing': True,
+            }
+        # soft_lims -> raise warning when outside lims
+        # hard_lim -> raise error when outside lims
+        # Format is [qMax, chiMax].
+        soft_param_lims = [20.01, 0.801]
+        hard_param_lims = [20.01, 0.99]
+        super(SEOBNRv4PHMSur, self).__init__(self.__class__.__name__, \
+            domain_type, keywords, soft_param_lims, hard_param_lims)
+
+    def _load_dimless_surrogate(self):
+        """
+        This function, which must be overriden for each derived class of
+        SurrogateEvaluator, handles the loading of the dimensionless surrogate.
+        This should return the loaded surrogate.
+        The loaded surrogate should have a __call__ function that returns the
+        dimensionless time/frequency array and dimensionless waveform modes.
+        The return value of this functions will be stored as
+        self._sur_dimless()
+        The __call__ function of self._sur_dimless() should take all inputs
+        passed to self._sur_dimless() in the __call__ function of this class.
+        See NRHybSur3dq8 for an example.
+        """
+
+        def get_fit_params(x):
+            """ convert user input x to parameters used by surrogate fits.
+            
+            Converts from x=[q, chi1x, chi1y, chi1z, chi2x, chi2y, chi2z]
+                      to  x=[q, chi1x, chi1y, chi1z, chi2x, chi2y, chi2z]
+            """
+            ## Literally do nothing!!
+            x = np.copy(x)
+            return x
+        
+        def get_fit_settings(subdomain_ID):
+            """ Get fit settings for SUBDOMAIN_ID.
+            
+            Arguments:
+                subdomain_ID: The subdomain ID. Each subdomain has a uniquely
+                            built surrogate model.
+            Returns:
+                q_fit_offset: The offset for the mass ratio (q) fit.
+                q_fit_slope: The slope for the mass ratio (q) fit.
[…]
