+++ .github/workflows/python-app.yml
@@ -29,7 +29,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install flake8 pytest nose wheel
+        pip install flake8 pytest wheel
         sudo apt-get -yq install gcc
         sudo apt-get -yq install libgsl-dev
 #    - name: Lint with flake8
