### 2024-08-22 sxs-collaboration/gwsurrogate#51 @sfield17
An ongoing PR (https://github.com/sxs-collaboration/gwsurrogate/pull/52) will temporarily restrict `numpy < 2.0` so at least installation works. Once this is merged in, we'll look into migrating to `numpy >= 2.0`
