### 2024-08-22 44613c6
Pass BH spins as vectors to get_subdomain
