### 2024-08-25 9b6716f
Add note about spin conventions in the orignal SEOBNRv4PHM model and surrogate.
