### 2024-08-25 45187bc
Add note about how the subdomain boundaries are defined
