### 2024-08-26 ad02492
ellMax=2 for now
