+++ gwsurrogate/new/precessing_surrogate.py
@@ -1005,7 +1005,7 @@ Returns:
         self._check_unused_opts(precessing_opts)
 
         if ellMax is None:
-            ellMax = 4
+            ellMax = 2      #FIXME
         if ellMax > 4:
             raise ValueError(f"NRSur7dq4 only allows ellMax<=4, got ellMax={ellMax}")
 
