### 2024-08-26 b6380e5
Update catalog for new Zenodo records
