+++ surrogate.py
@@ -1,0 +1,2538 @@
+""" Gravitational Wave Surrogate classes for text and hdf5 files"""
+
+from __future__ import division # for python 2
+
+__copyright__ = "Copyright (C) 2014 Scott Field and Chad Galley"
+__email__     = "sfield@umassd.edu, crgalley@tapir.caltech.edu"
+__status__    = "testing"
+__author__    = "Jonathan Blackman, Scott Field, Chad Galley, Vijay Varma, Kevin Barkett"
+__version__ = "1.1.6"
+__license__ = """
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in
+all copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+THE SOFTWARE.
+"""
+
+# adding "_" prefix to potentially unfamiliar module names
+# so they won't show up in gws' tab completion
+import numpy as np
+from scipy.interpolate import splrep as _splrep
+from scipy.interpolate import splev as _splev
+from gwtools.harmonics import sYlm as _sYlm
+from gwtools import plot_pretty as _plot_pretty
+from gwtools import gwtools as _gwtools # from the package gwtools, import the module gwtools (gwtools.py)....
+from gwtools import gwutils as _gwutils
+from .parametric_funcs import function_dict as my_funcs
+from .surrogateIO import H5Surrogate as _H5Surrogate
+from .surrogateIO import TextSurrogateRead as _TextSurrogateRead
+from .surrogateIO import TextSurrogateWrite as _TextSurrogateWrite
+from .surrogateIO import BHPTNRCalibValues as _BHPTNRCalibValues
+from gwsurrogate.new.surrogate import ParamDim, ParamSpace
+
+import warnings
+import os
+
+from .new import surrogate as new_surrogate
+from .new import precessing_surrogate
+from . import catalog
+from .catalog import get_modelID_from_filename
+
+try:
+  import matplotlib.pyplot as plt
+except:
+  print("Cannot load matplotlib.")
+
+try:
+  import h5py
+  h5py_enabled = True
+except ImportError:
+  h5py_enabled = False
+
+
+# needed to search for single mode surrogate directories
+def _list_folders(path,prefix):
+  '''returns all folders which begin with some prefix'''
+  import os as os
+  for f in os.listdir(path):
+    if f.startswith(prefix):
+      yield f
+
+# handy helper to save waveforms
+def write_waveform(t, hp, hc, filename='output',ext='bin'):
+  """write waveform to text or numpy binary file"""
+
+  if( ext == 'txt'):
+    np.savetxt(filename, [t, hp, hc])
+  elif( ext == 'bin'):
+    np.save(filename, [t, hp, hc])
[…]
