+++ gwsurrogate/surrogate.py
@@ -140,7 +140,7 @@ class EvaluateSingleModeSurrogate(_H5Surrogate, _TextSurrogateRead):
   #!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
   def __call__(self, q, M=None, dist=None, phi_ref=None,\
                      f_low=None, times=None, units='dimensionless',\
-                     singlemode_call=True):
+                     calibrated=True, singlemode_call=True):
     """Return single mode surrogate evaluation for...
 
        Input
@@ -155,7 +155,7 @@ class EvaluateSingleModeSurrogate(_H5Surrogate, _TextSurrogateRead):
        times           --- array of times at which surrogate is to be evaluated
        units           --- units (mks or dimensionless) of input array of time samples
        singlemode_call --- Never set this by hand, will be False if this routine is called by the multimode evaluator
-
+       calibrated      --- If set to false, the calibration coefficients alpha and beta are set to 1
 
        More information
        ================
@@ -183,7 +183,7 @@ class EvaluateSingleModeSurrogate(_H5Surrogate, _TextSurrogateRead):
 
     if (M is not None) and (dist is not None) and (times is not None) and (units != 'mks'):
     	raise ValueError('passing values of M, dist, and times suggest mks units should be used!')
-
+    
     if self.surrogate_mode_type == 'coorb_waveform_basis' and singlemode_call:
       msg = 'directly calling a coorb_waveform_basis surrogate will return a waveform in the co-orbital frame.\n'
       msg += 'Please use EvaluateSurrogate to evaluate co-orbital frame surrogates.\n'
@@ -196,18 +196,21 @@ class EvaluateSingleModeSurrogate(_H5Surrogate, _TextSurrogateRead):
     # at a given value of parameter space. These calibration parameters will scale the 
     # time and amplitude. Currently, these models use calibration:
     # BHPTNRSur1dq1e4, EMRISur1dq1e4
-    if(self.surrogateID == 'BHPTNRSur1dq1e4'):
+    if not calibrated:
+      alpha_nr_calibration, beta_nr_calibration = 1,1
+    elif(self.surrogateID == 'BHPTNRSur1dq1e4'):
       alpha_nr_calibration, beta_nr_calibration = self.compute_BHPT_calibration_params(q)
-    
-    if(self.surrogateID == 'EMRISur1dq1e4'):
+    elif(self.surrogateID == 'EMRISur1dq1e4'):
       # see Eq 4 of https://arxiv.org/abs/1910.10473
       # if alpha = 1 we have the "raw" waveform computed from
       # point-particle black hole perturbation theory
       # alpha rescales all modes in the same way in time and amplitude
       nu         = q/(1.+q)**2.
-      alpha_emri = 1.0-1.352854*nu-1.223006*nu*nu+8.601968*nu*nu*nu-46.74562*nu*nu*nu*nu
+      alpha_nr_calibration = 1.0-1.352854*nu-1.223006*nu*nu+8.601968*nu*nu*nu-46.74562*nu*nu*nu*nu
+      beta_nr_calibration = alpha_nr_calibration
     else:
-      alpha_emri = None
+      alpha_nr_calibration = None
+      beta_nr_calibration = None
 
 
     ### if (M,distance) provided, a physical mode in mks units is returned ###
@@ -220,9 +223,7 @@ class EvaluateSingleModeSurrogate(_H5Surrogate, _TextSurrogateRead):
       t_scale = 1.0
 
     # any model-specific amplitude scalings should go here
-    if(self.surrogateID == 'EMRISur1dq1e4'):
-      amp0 = alpha_emri * amp0
[…]
