### 2024-09-18 #7073 @plaplant
@Ceciliogq thanks for making progress on the review! Please feel free to make issues on the [upstream repository](https://github.com/sxs-collaboration/gwsurrogate) pointing out issues that should be addressed as part of this review. It would be excellent if any issues made reference this JOSS review thread. Let me know if you have questions about how to do this.

@GarethCabournDavies 👋 when you get a chance, please start making your way through the review issues. If you have questions about anything, I'm happy to help out!
