### 2024-09-20 #7073 @GarethCabournDavies
Note that in reviewing the conflict of interest statement, I am almost exactly one of the cases where a conflict of interest could be perceived, but is not an issue; many of the authors are also part of the LIGO scientific collaboration, and so have co-authored papers but we have not collaborated directly on any funded work
