### 2024-09-23 #7073 @GarethCabournDavies
## Notes on installation / basic usage
These are some comments on the installation / basic usage
with the tests that I wanted to note

from-source installation needs setuptools explicitly (I didn't install it straight away, and everything seemed fine until I import gwsurrogate in python, when things broke!)

Note that for all cases, I needed to go back and re-install with numpy specified as being <2, which I think is being dealt with in https://github.com/sxs-collaboration/gwsurrogate/issues/51

in pip and conda cases when first running import gwsurrogate I get a warning like:
```
/home/gareth/miniconda3/envs/joss_gwsurrogate_conda/lib/python3.12/site-packages/gwtools/const.py:52: UserWarning: Wswiglal-redir-stdio:

SWIGLAL standard output/error redirection is enabled in IPython.
This may lead to performance penalties. To disable locally, use:

with lal.no_swig_redirect_standard_output_error():
    ...

To disable globally, use:

lal.swig_redirect_standard_output_error(False)

Note however that this will likely lead to error messages from
LAL functions being either misdirected or lost when called from
Jupyter notebooks.

To suppress this warning, use:

import warnings
warnings.filterwarnings("ignore", "Wswiglal-redir-stdio")
import lal

  import lal
lal.MSUN_SI != Msun
```
Not an error, just a warning, thats fine, but would be better to work on how to get rid of it!

I am not a fan of the following being printed every time
```
__name__ = gwsurrogate.new.spline_evaluation
__package__= gwsurrogate.new
```
but if thats how you want to do it, then that's fine

Other minor problems:

Following the example in  https://github.com/sxs-collaboration/gwsurrogate/tree/master?tab=readme-ov-file#evaluate-the-surrogate, the dyn object returned is a None. This is at odds with the comment in the example about "do dyn.keys() to see contents"

In https://github.com/sxs-collaboration/gwsurrogate?tab=readme-ov-file#tests, need to add that pytest is required!
