### 2024-09-23 sxs-collaboration/gwsurrogate#54 @GarethCabournDavies
Error in readme / example usage
Using the example https://github.com/sxs-collaboration/gwsurrogate/tree/master?tab=readme-ov-file#download-surrogate-data-and-load-it

I found that once the surrogate has been evaluated, dyn is returned as None, this is at odds with the comment statement `dyn stands for dynamics, do dyn.keys() to see contents`

Looking into `help(sur)` for the NRSur7dq4 surrogate, it says `This is None for nonprecessing models`, so I think this is just a case of updating that comment
