### 2024-09-23 sxs-collaboration/gwsurrogate#55 @GarethCabournDavies
Tests minor addition
Minor request - https://github.com/sxs-collaboration/gwsurrogate/tree/master?tab=readme-ov-file#tests should say that pytest need installing!
