### 2024-09-23 #7073 @plaplant
@GarethCabournDavies thanks for your review! I appreciate your acknowledging your affiliation with LIGO. I agree that this does not rise to the level of a conflict of interest, so I am formally waiving this as a potential issue. I also appreciate your making issues on the upstream repository pointing out some potential improvements. This is helpful for bookkeeping the review issues.

@sfield17 both reviewers have taken an initial pass at the software package, and pointed out some things that should be addressed. When you get a chance, please start responding to these issues (either in this thread or in the issues on the main repository, whichever you deem more appropriate). When these issues have been resolved, please respond to this review thread saying so. At that point the reviewers can take another pass to see if they have been addressed to their satisfaction.

Thanks everyone for contributing to this review process! Please let me know if you have any questions.
