### 2024-09-23 89b5a54
Revert Zenodo record's URL
The reverted URL points to the same hdf5 model file (ie same hash), but the
older record preserves information about which record version we were on
when the model's hdf5 file was last updated.
