### 2025-01-06 12790b3
Revert "ellMax=2 for now"
This reverts commit ad02492c84b4eb0cad95c8bc398ee1f0baaa400a.
