### 2025-01-07 f3a6dc6
ellMax=5
