### 2025-01-10 5ffc72b
Remove unnecessary code
