### 2025-01-10 d52e77b
Simplify variable naming convention
