### 2025-01-12 ef3e46a
Simplify code to check if hdf5 group exists
