### 2025-01-13 c7be56b
Use raw strings when docstring contains a backslash. See PEP 257
