+++ gwsurrogate/surrogate.py
@@ -306,7 +306,7 @@ class EvaluateSingleModeSurrogate(_H5Surrogate, _TextSurrogateRead):
 
   #!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
   def find_instant_freq(self, hp, hc, t, f_low = None):
-    """instantaneous frequency at t_start for
+    r"""instantaneous frequency at t_start for
 
           h = A(t) exp(2 * pi * i * f(t) * t),
 
@@ -1724,7 +1724,7 @@ class SurrogateEvaluator(object):
         precessing_opts=None, tidal_opts=None, par_dict=None,
         units='dimensionless', skip_param_checks=False,
         taper_end_duration=None):
-        """
+        r"""
     INPUT
     =====
     q :         Mass ratio, mA/mB >= 1.
@@ -2133,7 +2133,7 @@ class SurrogateEvaluator(object):
 
 
 class NRHybSur3dq8(SurrogateEvaluator):
-    """
+    r"""
 A class for the NRHybSur3dq8 surrogate model presented in Varma et al. 2018,
 arxiv:1812.07865.
 
@@ -2205,9 +2205,9 @@ See the __call__ method on how to evaluate waveforms.
 
 
 class NRHybSur3dq8Tidal(SurrogateEvaluator):
-    """
+    r"""
 A class for the NRHybSur3dq8Tidal model presented in Barkett et al.,
-arxiv:xxxx.xxxx #FIXME.
+arxiv:1911.10440. 
 
 Generates inspiralling gravitational waveforms corresponding to binary neutron
 stars/black hole-neutron star systems. This model is based on the aligned-spin
@@ -2290,7 +2290,7 @@ See the __call__ method on how to evaluate waveforms.
         return x
 
 class NRHybSur3dq8_CCE(NRHybSur3dq8):
-    """
+    r"""
 A class for the NRHybSur3dq8_CCE surrogate model presented in arXiv:2306.03148.
 
 Evaluates gravitational waveforms generated by aligned-spin binary black hole
@@ -2319,7 +2319,7 @@ See the __call__ method on how to evaluate waveforms.
     pass
 
 class NRHybSur2dq15(SurrogateEvaluator):
-    """
+    r"""
 A class for the NRHybSur2dq15 surrogate model presented in
 arxiv:2203.10109. 
 
@@ -2391,7 +2391,7 @@ See the __call__ method on how to evaluate waveforms.
         return x
 
 class NRSur7dq4(SurrogateEvaluator):
-    """
+    r"""
 A class for the NRSur7dq4 surrogate model presented in Varma et al. 2019,
 arxiv1905.09300.
 
@@ -2507,7 +2507,7 @@ See the __call__ method on how to evaluate waveforms.
         return x
 
 class SEOBNRv4PHMSur(SurrogateEvaluator):
-    """
+    r"""
 A class for the SEOBNRv4PHM surrogate models.
 
 Evaluates gravitational waveforms generated by precessing binary black hole
