### 2025-01-13 5c85629
Add model regression test for SEOBNRv4PHMSur
