### 2025-01-14 5445e91
Remove system-dependent wget in favor of the Python requests package.
This change addresses two open issues where users were unable to
download surrogates using wget on their machine.
