+++ gwsurrogate/catalog.py
@@ -309,6 +309,7 @@ def pull(surr_name,sdir=download_path()):
                 ' some.'%backup_dir)
 
     # download the surrogate
+    os.makedirs(sdir, exist_ok=True) # Ensure the target directory exists (mimicking wget's --directory-prefix functionality)
     filename = surr_url.split("/")[-1]
     output_path = os.path.join(sdir, filename)
     with requests.get(surr_url, stream=True) as r, open(output_path, "wb") as f:
