### 2025-01-16 845777b
Add pytest to installation requirements
