### 2025-01-17 7dd76a2
Clarified READMEs example surrogate usage for the NRSur7dq4 model.
