### 2025-01-17 98df0af
Add pytest to list of installation requirements
