### 2025-01-17 998d407
Add Contributing and Documentation sections to README
