+++ tutorial/paper/paper.bib
@@ -162,3 +162,167 @@
     pages = "104025",
     year = "2022"
 }
+
+@article{Gadre:2022sed,
+    author = {Gadre, Bhooshan and P\"urrer, Michael and Field, Scott E. and Ossokine, Serguei and Varma, Vijay},
+    title = "{Fully precessing higher-mode surrogate model of effective-one-body waveforms}",
+    eprint = "2203.00381",
+    archivePrefix = "arXiv",
+    primaryClass = "gr-qc",
+    reportNumber = "LIGO-P2200040",
+    doi = "10.1103/PhysRevD.110.124038",
+    journal = "Phys. Rev. D",
+    volume = "110",
+    number = "12",
+    pages = "124038",
+    year = "2024"
+}
+
+@article{Rink:2024swg,
+    author = "Rink, Katie and Bachhar, Ritesh and Islam, Tousif and Rifat, Nur E. M. and Gonzalez-Quesada, Kevin and Field, Scott E. and Khanna, Gaurav and Hughes, Scott A. and Varma, Vijay",
+    title = "{Gravitational wave surrogate model for spinning, intermediate mass ratio binaries based on perturbation theory and numerical relativity}",
+    eprint = "2407.18319",
+    archivePrefix = "arXiv",
+    primaryClass = "gr-qc",
+    doi = "10.1103/PhysRevD.110.124069",
+    journal = "Phys. Rev. D",
+    volume = "110",
+    number = "12",
+    pages = "124069",
+    year = "2024"
+}
+
+@article{Islam:2021mha,
+    author = "Islam, Tousif and Varma, Vijay and Lodman, Jackie and Field, Scott E. and Khanna, Gaurav and Scheel, Mark A. and Pfeiffer, Harald P. and Gerosa, Davide and Kidder, Lawrence E.",
+    title = "{Eccentric binary black hole surrogate models for the gravitational waveform and remnant properties: comparable mass, nonspinning case}",
+    eprint = "2101.11798",
+    archivePrefix = "arXiv",
+    primaryClass = "gr-qc",
+    doi = "10.1103/PhysRevD.103.064022",
+    journal = "Phys. Rev. D",
+    volume = "103",
+    number = "6",
+    pages = "064022",
+    year = "2021"
+}
+
+ @misc{lalsuite,
+       author         = "{LIGO Scientific Collaboration} and {Virgo Collaboration} and {KAGRA Collaboration}",
+       title          = "{LVK} {A}lgorithm {L}ibrary - {LALS}uite",
+       howpublished   = "Free software (GPL)",
+       doi            = "10.7935/GT1W-FZ16",
+       year           = "2018"
+ }
+
+@article{Mihaylov:2023bkc,
+  author = {Mihaylov, Deyan P. and Ossokine, Serguei and Buonanno, Alessandra and Estelles, Hector and Pompili, Lorenzo and P\"urrer, Michael and Ramos-Buades, Antoni},
+  title = "{pySEOBNR: a software package for the next generation of effective-one-body multipolar waveform models}",
+  eprint = "2303.18203",
+  archivePrefix = "arXiv",
+  primaryClass = "gr-qc",
+  month = "3",
+  year = "2023"
+}
+
+@article{Nagar:2020pcj,
+    author = "Nagar, Alessandro and Riemenschneider, Gunnar and Pratten, Geraint and Rettegno, Piero and Messina, Francesco",
+    title = "{Multipolar effective one body waveform model for spin-aligned black hole binaries}",
+    eprint = "2001.09082",
+    archivePrefix = "arXiv",
+    primaryClass = "gr-qc",
+    doi = "10.1103/PhysRevD.102.024077",
[…]
