+++ tutorial/paper/paper.md
@@ -99,7 +99,9 @@ where $^{-2}Y_{\ell m}$ are the spin$=-2$ weighted spherical harmonics and $\Lam
 - models implemented in ``GWSurrogate`` follow the waveform convention choices of the LIGO-Virgo-Kagra collaboration, thus ensuring that downstream data analysis codes can use ``GWSurrogate`` models without needing to worry about different conventions, and
 - ``GWSurrogate`` models can be directly evaluated in either physical units (often used in data analysis studies) and dimensionless units (often used in theoretical studies) where all dimensioned quantities are expressed in terms of the system's total mass.
 
-Currently, there are 15 supported surrogate models (@Blackman:2015pia, @OShaughnessy:2017tak, @Blackman:2017dfb, @Varma:2018mmi, @Yoo:2023spi, @Yoo:2022erv, @Varma:2019csw, @Barkett:2019tus, @Rifat:2019ltp, @Islam:2022laz, @Field:2013cfa). These models vary in their duration, included physical effects (e.g. nonlinear memory, tidal forces, harmonic modes retained, eccentricity, mass ratio extent, precession effects, etc), and underlying solution method (e.g. Effective One Body, numerical relativity, and black hole perturbation theory). Details about all models can be found by doing `gwsurrogate.catalog.list(verbose=True)`, while the ``GWSurrogate`` [homepage](https://github.com/sxs-collaboration/gwsurrogate) summarizes the state-of-the-art models for each particular problem. Certain models allow for additional functionality such as returning the dynamics of the binary black hole; these special features are described further in model-specific [example notebooks](https://github.com/sxs-collaboration/gwsurrogate/tree/master/tutorial).
+Currently, there are 16 supported surrogate models (@Blackman:2015pia, @OShaughnessy:2017tak, @Blackman:2017dfb, @Varma:2018mmi, @Yoo:2023spi, @Yoo:2022erv, @Varma:2019csw, @Barkett:2019tus, @Rifat:2019ltp, @Islam:2022laz, @Field:2013cfa, @Gadre:2022sed), with additional models under development (@Rink:2024swg, @Islam:2021mha). These models vary in their duration, included physical effects (e.g. nonlinear memory, tidal forces, harmonic modes retained, eccentricity, mass ratio extent, precession effects, etc), and underlying solution method (e.g. Effective One Body, numerical relativity, and black hole perturbation theory). Details about all models can be found by doing `gwsurrogate.catalog.list(verbose=True)`, while the ``GWSurrogate`` [homepage](https://github.com/sxs-collaboration/gwsurrogate) summarizes the state-of-the-art models for each particular problem. Certain models allow for additional functionality such as returning the dynamics of the binary black hole; these special features are described further in model-specific [example notebooks](https://github.com/sxs-collaboration/gwsurrogate/tree/master/tutorial).
+
[…]
