### 2025-01-17 959e391
Update draft-pdf.yml
Use v4 of artifact actions as suggested by GitHub.
