+++ tutorial/paper/paper.md
@@ -101,7 +101,7 @@ where $^{-2}Y_{\ell m}$ are the spin$=-2$ weighted spherical harmonics and $\Lam
 
 Currently, there are 16 supported surrogate models (@Blackman:2015pia, @OShaughnessy:2017tak, @Blackman:2017dfb, @Varma:2018mmi, @Yoo:2023spi, @Yoo:2022erv, @Varma:2019csw, @Barkett:2019tus, @Rifat:2019ltp, @Islam:2022laz, @Field:2013cfa, @Gadre:2022sed), with additional models under development (@Rink:2024swg, @Islam:2021mha). These models vary in their duration, included physical effects (e.g. nonlinear memory, tidal forces, harmonic modes retained, eccentricity, mass ratio extent, precession effects, etc), and underlying solution method (e.g. Effective One Body, numerical relativity, and black hole perturbation theory). Details about all models can be found by doing `gwsurrogate.catalog.list(verbose=True)`, while the ``GWSurrogate`` [homepage](https://github.com/sxs-collaboration/gwsurrogate) summarizes the state-of-the-art models for each particular problem. Certain models allow for additional functionality such as returning the dynamics of the binary black hole; these special features are described further in model-specific [example notebooks](https://github.com/sxs-collaboration/gwsurrogate/tree/master/tutorial).
 
-Several other software packages are available for waveform generation, including tools for effective-one-body models (@Mihaylov:2023bkc, @Nagar:2020pcj), ringdown signals (@pyRing, @Isi:2021iql, @alex_nitz_2024_10473621), extreme-mass-ratio inspiral systems (@BHPToolkit) (notably the ``FastEMRIWaveforms`` and ``BHPTNRSurrogate`` packages), and the ``Ripple`` framework that enables specialized acceleration techniques using JAX (@Edwards:2023sak). Among these, LALSuite (@lalsuite) stands out as the most comprehensive, offering the largest collection of waveform models via its LALSimulation subpackage, which includes Python bindings and the new Python-based gwsignal waveform generator. While GWSurrogate shares similarities with LALSuite in providing a variety of models, it differs by exclusively focusing on surrogate models. Notably, GWSurrogate includes many state-of-the-art numerical relativity models that are only available through its library, whereas LALSuite offers a broader but less specialized collection.
+Several other software packages are available for waveform generation, including tools for effective-one-body models (@Mihaylov:2023bkc, @Nagar:2020pcj), ringdown signals (@pyRing, @Isi:2021iql, @alex_nitz_2024_10473621), extreme-mass-ratio inspiral systems through the Black Hole Perturbation Toolkit's ``FastEMRIWaveforms`` and ``BHPTNRSurrogate`` packages (@BHPToolkit), and the ``Ripple`` framework that enables specialized acceleration techniques using JAX (@Edwards:2023sak). Among these, LALSuite (@lalsuite) stands out as the most comprehensive, offering the largest collection of waveform models via its LALSimulation subpackage, which includes Python bindings and the new Python-based gwsignal waveform generator. While GWSurrogate shares similarities with LALSuite in providing a variety of models, it differs by exclusively focusing on surrogate models. Notably, GWSurrogate includes many state-of-the-art numerical relativity models that are only available through its library, whereas LALSuite offers a broader but less specialized collection.
 
 
 # Acknowledgements
