### 2025-01-17 ebae0be
Remove outdated README instructions for installing from source via tar.gz file
