### 2025-01-18 11d05df
Remove unnecessary print statements
