### 2025-01-18 #7073 @sfield17
@plaplant We have summarized all changes to the code and paper in response to the reviewers' feedback.

## Overview of Changes

We thank @GarethCabournDavies and @Ceciliogq  for their time and feedback on our paper and code. We also apologize for the delayed response. Part of the delay was due to our efforts to port another model to GWSurrogate, as part of a publication that committed to making the model publicly available ([Phys. Rev. D 110, 124038](https://journals.aps.org/prd/abstract/10.1103/PhysRevD.110.124038)). Below, we summarize the changes related to this new model:

- **Implementation of the model**: [Pull Request #52](https://github.com/sxs-collaboration/gwsurrogate/pull/52)
- **Regression test for the new model**: [Pull Request #56](https://github.com/sxs-collaboration/gwsurrogate/pull/56)

Both of these changes have been merged into GWSurrogate's main branch. In addition to these updates, we provide a summary of changes made during the ongoing JOSS paper review. These include updates to both the code and the manuscript, followed by specific responses to each reviewer's comments.

### **Code Changes**
The edits made as part of the JOSS paper review are included in the `Joss_paper_review` branch and the corresponding [Pull Request #58](https://github.com/sxs-collaboration/gwsurrogate/pull/58). The following changes have been implemented:

- Addressed an issue with the README example ([Issue #54](https://github.com/sxs-collaboration/gwsurrogate/issues/54)).
- Resolved an issue with the pytest installation requirement ([Issue #55](https://github.com/sxs-collaboration/gwsurrogate/issues/55)).
- Added a [Documentation](https://github.com/sxs-collaboration/gwsurrogate/tree/joss_paper_review?tab=readme-ov-file#documentation) section to the README file, which consolidates key information and includes links to talks where GWSurrogate has been discussed.
- Added a [Contributing](https://github.com/sxs-collaboration/gwsurrogate/tree/joss_paper_review?tab=readme-ov-file#contributing) section to the README file. This section provides contribution guidelines, instructions for using the issue tracker, and support information.
- Removed the superfluous print statements, resulting in the following being printed every time gwsurrogate is imported

```
__name__ = gwsurrogate.new.spline_evaluation
__package__= gwsurrogate.new
```
- Removed the "From source (tar.gz)" installation option from the README, as it results in a broken installation and is unnecessary. The updated installation notes are [here](https://github.com/sxs-collaboration/gwsurrogate/tree/joss_paper_review?tab=readme-ov-file#installation)

### **Paper Changes**
We have made the following updates to the paper:

1. These changes reflect recent updates in the GWSurrogate codebase over the past few months:
   - Added a reference to ([Phys. Rev. D 110, 124038](https://journals.aps.org/prd/abstract/10.1103/PhysRevD.110.124038)), which corresponds to the 16th model included in the package.
   - Included details about additional models currently under development.

2.  In response to the reviewers’ feedback, we added a new paragraph to the manuscript, beginning with:
   *"Several other software packages are available for waveform generation..."*
   This paragraph summarizes other waveform generation software and clarifies how GWSurrogate compares.

3. Reference 2013 as the start of the GWSurrogate package.

## Reply to @GarethCabournDavies

#### **Data Sharing**

Regarding your comment:
*"The data behind the figure is available at https://github.com/sxs-collaboration/gwsurrogate/blob/joss_paper/tutorial/paper/figs/data.npy, though how that is gained is not explained. As with @Ceciliogq, I don't think this is necessarily original data, and so not a problem."*

The figure is intended to visually illustrate a typical waveform generated by the GWSurrogate code (in this case, from a precessing surrogate model) and to provide computational cost numbers for the original simulation and surrogate evaluation. It is not meant to be exactly reproducible but serves as an engaging visual for the paper.

If exact reproducibility is required, we could substitute the image with one from another publication, where users could consult the original reference for details. However, we prefer to retain the current image, as we believe additional reproducibility here would not add significant value.

Note: @Ceciliogq raised a similar question regarding data sharing.

#### **Original Results**

We confirm that the paper does not contain any original scientific results.

#### **Functionality Documentation**

Regarding your comment:
*"There is no API method documentation that I can see - this means that either it doesn't exist, or it is too hard to find!"*

The user-level API is primarily documented through [tutorial Jupyter notebooks](https://github.com/sxs-collaboration/gwsurrogate/tree/master/tutorial/website). These are linked from the README file at [this section](https://github.com/sxs-collaboration/gwsurrogate?tab=readme-ov-file#current-nr-models) and [here](https://github.com/sxs-collaboration/gwsurrogate?tab=readme-ov-file#evaluate-the-surrogate).

To make this information easier to locate, we have added a [Documentation](https://github.com/sxs-collaboration/gwsurrogate/tree/joss_paper_review?tab=readme-ov-file#documentation) section to the README file. This new section also includes links to talks discussing GWSurrogate.

#### **Community Guidelines**

Regarding your comment:
*"There is a (extensive, which I commend) code of conduct. I do not see contribution guidelines. Issues are as standard for GitHub, but guidelines on how to use the issues (e.g., issue templates) are not given. Support is not explained."*

To address these suggestions, we have added a new [Contributing](https://github.com/sxs-collaboration/gwsurrogate/tree/joss_paper_review?tab=readme-ov-file#contributing) section to the README file. This section outlines contribution guidelines, instructions for using the issue tracker, and support resources.

#### **Comparison to the State of the Field**

Regarding your comment:
*"Please compare to other software, particularly the Phenom and SEOB families of waveforms (and their availability)."*

We have added a new paragraph to the manuscript, beginning with:
*"Several other software packages are available for waveform generation..."*
This paragraph provides a summary of other waveform generation software and explains how GWSurrogate compares.

#### **Notes on Installation / Basic Usage**

In response to your feedback, we have removed superfluous print statements and addressed the two issues opened in the GWSurrogate issue tracker

## Reply to @Ceciliogq

#### **Data Sharing**

Regarding your comment:
*"I am not sure what original data means. They show a plot of a Numerical Relativity waveform which is not available, but I do not think that should be a problem."*

We fully agree. The figure is intended to visually illustrate a typical waveform generated by the GWSurrogate code (from a precessing surrogate model) and provide computational cost numbers for the original simulation and surrogate evaluation. It is not meant to be exactly reproducible but serves as an engaging visual for the paper.

If exact reproducibility is required, we could substitute the image with one from another publication, where users could consult the original reference for details. However, we prefer to retain the current image, as we believe additional reproducibility would not add significant value.

Note: GarethCabournDavies raised a similar question regarding data sharing.

#### **Reproducibility**

Regarding your comment:
*"If by results you mean the plot comparing the surrogate with the simulation, then there is not a clear way to reproduce it since I do not know the parameters of that simulation. However, I think there is no need to explicitly check this."*

We agree that explicit checks are unnecessary, given the intention of the figure, as explained above.

#### **Installation**

Regarding your comment:
*"Installing from pip and conda was fine. However, I did not manage to install from source. When importing gwsurrogate I get..."*

The README describes two "Install from source" methods. The version [From source (pip)](https://github.com/sxs-collaboration/gwsurrogate?tab=readme-ov-file#from-source-pip) works for us (notably, the two git submodule commands must be run). However, the other option, [From source (tar.gz)](https://github.com/sxs-collaboration/gwsurrogate?tab=readme-ov-file#from-source-targz), fails due to the missing module `gwsurrogate.eval_pysur`.

We have removed the "From source (tar.gz)" option from the README, as it results in a broken installation and is unnecessary. The updated installation notes are [here](https://github.com/sxs-collaboration/gwsurrogate/tree/joss_paper_review?tab=readme-ov-file#installation)

#### **Automated Tests**

Regarding your comment:
*"There are automated tests and instructions how to run them, but I did not manage to make it work. When following the instructions I get..."*

The `ImportError` suggests the installation process did not include the git submodule commands mentioned [here](https://github.com/sxs-collaboration/gwsurrogate?tab=readme-ov-file#from-source-pip). This issue likely arose from following the broken "From source (tar.gz)" instructions, which have since been removed in the [Joss_paper_review branch](https://github.com/sxs-collaboration/gwsurrogate/pull/58).

Following the newly clarified installation instructions [From source (pip)](https://github.com/sxs-collaboration/gwsurrogate/tree/joss_paper_review?tab=readme-ov-file#from-source-pip) should allow the tests to run locally without errors.

#### **Community Guidelines**

Regarding your comment:
*"I have not seen these."*

We have added a new [Contributing](https://github.com/sxs-collaboration/gwsurrogate/tree/joss_paper_review?tab=readme-ov-file#contributing) section to the README file, which outlines contribution guidelines, instructions for using the issue tracker, and support information.

Our Code of Conduct is also referenced [here](https://github.com/sxs-collaboration/gwsurrogate?tab=coc-ov-file).

#### **State of the Field**

Regarding your comment:
*"I do not see this."*

We have added a new paragraph to the manuscript, beginning with:
*"Several other software packages are available for waveform generation..."*
This paragraph summarizes other waveform generation software and explains how GWSurrogate compares.
