### 2025-02-11 #7073 @plaplant
@GarethCabournDavies @Ceciliogq 👋 thanks again for your initial reviews! When you get a chance, please take a look at the [author's response to the issues raised](https://github.com/openjournals/joss-reviews/issues/7073#issuecomment-2599793247). Feel free to raise more issues or update your reviewer checklists for completed items, whichever is most appropriate. Let me know if you have any questions!
