### 2025-02-13 #7073 @Ceciliogq
Thanks @sfield17. I am also happy with the responses. I only opened a minor issue about the automated tests.

Also, where can I find the latest version of the paper with the changes @plaplant ?
