### 2025-02-13 #7073 @sfield17
Thanks @GarethCabournDavies and @Ceciliogq for the feedback and suggestions. I'll take a look at the minor issues caught by @Ceciliogq.

@Ceciliogq The most recent version of the paper (built from pushes to the joss_paper branch) [can be found here](https://github.com/sxs-collaboration/gwsurrogate/actions/runs/12845861776). At the bottom of this action is something called "artifacts". If you click on "paper" it should download the paper.
