### 2025-02-13 4409732
Update discussion of tests in readme
