### 2025-02-13 3efea2e
Update discussion of tests in readme
