### 2025-02-23 #7073 @plaplant
@GarethCabournDavies thanks for your response! I've gone ahead and ticked all of the boxes in your reviewer checklist to reflect the fact that you're satisfied with all of the changes.

@Ceciliogq thanks for responding as well! In addition to the link provided by @sfield17, you can also see the latest version of the paper [here](https://github.com/openjournals/joss-reviews/issues/7073#issuecomment-2676417945). When you feel the authors have sufficiently responded to the outstanding issues, please go ahead and update your user checklist. If the issues have not been addressed or require additional attention, please feel free to open additional issues.

Thanks everyone! Let me know if you have any questions.
