### 2025-02-26 sxs-collaboration/gwsurrogate#54 @sfield17
This was fixed as part of the JOSS paper review referenced above. The paper review branch has been merged into the main branch of gwsurrogate as part of https://github.com/sxs-collaboration/gwsurrogate/pull/58
