### 2025-02-26 9ad8034
Add CITATION.cff
This data is currently static. Either update it for each version or
any time author info changes; or use a github action to update it.
