+++ CITATION.cff
@@ -1,0 +1,42 @@
+cff-version: 1.2.0
+message: "If you use this software, please cite it as below."
+authors:
+  - family-names: "Field"
+    given-names: "Scott E."
+    orcid: "https://orcid.org/0000-0002-6037-3277"
+    equal-contrib: true
+  - family-names: "Varma"
+    given-names: "Vijay"
+    orcid: "https://orcid.org/0000-0002-9994-1761"
+    equal-contrib: true
+  - family-names: "Blackman"
+    given-names: "Jonathan"
+  - family-names: "Gadre"
+    given-names: "Bhooshan"
+    orcid: "https://orcid.org/0000-0002-1534-9761"
+  - family-names: "Galley"
+    given-names: "Chad R."
+  - family-names: "Islam"
+    given-names: "Tousif"
+    orcid: "https://orcid.org/0000-0002-3434-0084"
+  - family-names: "Mitman"
+    given-names: "Keefe"
+    orcid: "https://orcid.org/0000-0003-0276-3856"
+  - family-names: "Pürrer"
+    given-names: "Michael"
+    orcid: "https://orcid.org/0000-0002-3329-9788"
+  - family-names: "Ravichandran"
+    given-names: "Adhrit"
+  - family-names: "Scheel"
+    given-names: "Mark A."
+    orcid: "https://orcid.org/0000-0001-6656-9134"
+  - family-names: "Stein"
+    given-names: "Leo C."
+    orcid: "https://orcid.org/0000-0001-7559-9597"
+  - family-names: "Yoo"
+    given-names: "Jooheon"
+    orcid: "https://orcid.org/0000-0002-3251-0924"
+title: "gwsurrogate"
+version: 1.1.7
+date-released: 2025-02-25
+url: "https://github.com/sxs-collaboration/gwsurrogate"
