### 2025-02-26 9e02395
Add license to CITATION.cff; use `repository-code` instead of `url`
