### 2025-03-01 #7073 @sfield17
@plaplant I believe we have done all of the above. In particular...

1. Double-check authors and affiliations (including ORCIDs)

- A few minor changes were made. This is in the latest push to the [paper branch](https://github.com/sxs-collaboration/gwsurrogate/tree/joss_paper)

2. Make a release of the software with the latest changes from the review and post the version number

- Merged the JOSS paper review branch into the main branch
- Updated the version to **1.1.7**.
- Made a GitHub release for version **1.1.7**, which is also on pypi [gwsurrogate 1.1.7](https://pypi.org/project/gwsurrogate/1.1.7/)

3. Archive the release on Zenodo/figshare and post the DOI

- Zenodo DOI is: [10.5281/zenodo.14933935](https://zenodo.org/records/14933936)

4. Ensure that the title and author list (including ORCIDs) in the archive match the JOSS paper

- Done.

5. Verify that the archive license matches the software license

- Done.
