### 2025-03-19 #7073 @plaplant
@sfield17 thanks for your patience, I have only just been able to get around to reviewing this. Very nice work on the submission! I only had a few small DOI/journal touch-ups to add as part of [a PR I made on the repo](https://github.com/sxs-collaboration/gwsurrogate/pull/61). Feel free to review and merge when you get a chance, and we should be able to move forward with acceptance.

Thanks! Let me know if you have any questions.
