### 2025-03-22 #7073 @sfield17
@plaplant Thanks for opening up the very helpful PR. I've just [merged this](https://github.com/sxs-collaboration/gwsurrogate/pull/61) and the paper compiles just fine, which now includes fixes to the DOI issues flagged above.
