### 2025-03-23 #7073 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
