### 2025-03-23 #7073 @sfield17
@plaplant the final proof looks good to me. I wasn't sure if I was supposed to run the `@editorialbot accept` command. Or if thats something for the editor to do.
