### 2025-03-23 #7073 @plaplant
@sfield17 the final `accept` command can only be run by one of JOSS's editors-in-chief. This usually happens within a few days, but I will follow up if there hasn't been activity by next week.
