### 2025-03-23 #7073 @warrickball
Hi @sfield17, I'm the associate editor-in-chief (AEiC) for the astro track and it'll be up to me to run the `@editorialbot accept` command. (Don't worry: if you try you'd just only get an error message anyway.) Thanks for taking a look at the proofs. There are a few papers in the publication queue at the moment so I'll probably get around to making a pull requests with some final tweaks towards the end of this week.
