### 2025-03-24 #7073 @warrickball
Hi @sfield17, could you update the title in the Zenodo archive to match the JOSS paper? You don't need to create a new archive; you should be able to change the existing metadata in the Zenodo UI.
