### 2025-03-28 a773712
Initial editorial tweaks for JOSS paper
