+++ tutorial/paper/paper.bib
@@ -1,7 +1,7 @@
 @article{LISAConsortiumWaveformWorkingGroup:2023arg,
-    author = "Afshordi, Niayesh and others",
+    author = {{Afshordi}, Niayesh and {Ak{\c{c}}ay}, Sarp and {Amaro Seoane}, Pau and {Antonelli}, Andrea and {Aurrekoetxea}, Josu C. and {Barack}, Leor and {Barausse}, Enrico and {Benkel}, Robert and {Bernard}, Laura and {Bernuzzi}, Sebastiano and {Berti}, Emanuele and {Bonetti}, Matteo and {Bonga}, B{\'e}atrice and {Bozzola}, Gabriele and {Brito}, Richard and {Buonanno}, Alessandra and {C{\'a}rdenas-Avenda{\~n}o}, Alejandro and {Casals}, Marc and {Chernoff}, David F. and {Chua}, Alvin J.~K. and {Clough}, Katy and {Colleoni}, Marta and {Dhesi}, Mekhi and {Druart}, Adrien and {Durkan}, Leanne and {Faye}, Guillaume and {Ferguson}, Deborah and {Field}, Scott E. and {Gabella}, William E. and {Garc{\'\i}a-Bellido}, Juan and {Gracia-Linares}, Miguel and {Gerosa}, Davide and {Green}, Stephen R. and {Haney}, Maria and {Hannam}, Mark and {Heffernan}, Anna and {Hinderer}, Tanja and {Helfer}, Thomas and {Hughes}, Scott A. and {Husa}, Sascha and {Isoyama}, Soichiro and {Katz}, Michael L. and {Kavanagh}, Chris and {Khanna}, Gaurav and {Kidder}, Larry E. and {Korol}, Valeriya and {K{\"u}chler}, Lorenzo and {Laguna}, Pablo and {Larrouturou}, Fran{\c{c}}ois and {Le Tiec}, Alexandre and {Leather}, Benjamin and {Lim}, Eugene A. and {Lim}, Hyun and {Littenberg}, Tyson B. and {Long}, Oliver and {Lousto}, Carlos O. and {Lovelace}, Geoffrey and {Lukes-Gerakopoulos}, Georgios and {Lynch}, Philip and {Macedo}, Rodrigo P. and {Markakis}, Charalampos and {Maggio}, Elisa and {Mandel}, Ilya and {Maselli}, Andrea and {Mathews}, Josh and {Mourier}, Pierre and {Neilsen}, David and {Nagar}, Alessandro and {Nichols}, David A. and {Nov{\'a}k}, Jan and {Okounkova}, Maria and {O'Shaughnessy}, Richard and {Oshita}, Naritaka and {O'Toole}, Conor and {Pan}, Zhen and {Pani}, Paolo and {Pappas}, George and {Paschalidis}, Vasileios and {Pfeiffer}, Harald P. and {Pompili}, Lorenzo and {Pound}, Adam and {Pratten}, Geraint and {R{\"u}ter}, Hannes R. and {Ruiz}, Milton and {Sam}, Zeyd and {Sberna}, Laura and {Shapiro}, Stuart L. and {Shoemaker}, Deirdre M. and {Sopuerta}, Carlos F. and {Spiers}, Andrew and {Sundar}, Hari and {Tamanini}, Nicola and {Thompson}, Jonathan E. and {Toubiana}, Alexandre and {Tsokaros}, Antonios and {Upton}, Samuel D. and {van de Meent}, Maarten and {Vernieri}, Daniele and {Wachter}, Jeremy M. and {Warburton}, Niels and {Wardell}, Barry and {Witek}, Helvi and {Witzany}, Vojt{\v{e}}ch and {Yang},
[…]
