### 2025-03-28 #7073 @warrickball
I've opened [a pull request](https://github.com/sxs-collaboration/gwsurrogate/pull/62) with an initial set of editorial tweaks. If you merge those (or the subset you're happy with) I'll do another test run of the publication pipeline and check the metadata deposits again.
