### 2025-03-29 #7073 @sfield17
@warrickball I've merged in this PR. Thanks for making those changes!

Upon rereading the paper, I also noticed one very minor (but potentially awkward) phrasing regarding the distinction between HDF5 files and the model itself. I've fixed this as part of the following commit:

https://github.com/sxs-collaboration/gwsurrogate/commit/a0acc27af20f3186ee5d51987f99fcb0eb0323e8

Despite it being an extremely minor change, I wanted to make sure the editors were aware of it.
