### 2025-03-29 #7073 @warrickball
Congratulations @sfield17, your paper has been published in JOSS! 🎉
